Reimplement Treecapitator
